requirements.txt
Flask==1.1.2
pandas==1.0.5
psycopg2==2.8.5
pymongo==3.10.1
PyJWT==1.7.1
SQLAlchemy==1.3.18
dnspython==2.0.0
